Adds tests for Movie Item ui

+++ app/src/main/java/com/juawapps/whatstowatch/movies/data/model/MovieListItemDTO.kt
@@ -11,7 +11,7 @@ data class MovieListItemDTO(
     @Json(name = "overview") val overview: String,
     @Json(name = "release_date") val releaseDate: Date,
     @Json(name = "genre_ids") val genreIds: List<Int>,
-    @Json(name = "id") val id: Int,
+    @Json(name = "id") val id: Long,
     @Json(name = "original_title") val originalTitle: String,
     @Json(name = "original_language") val originalLanguage: String,
     @Json(name = "title") val title: String,
